pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
authors = [{ name = "Alessio Arcara" }]
name = "rtmdet"
version = "0.3.0"
description = "RTMDet Pytorch Implementation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "numpy<2.0",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/alessioarcara/rtmdet_pytorch_implementation"
Repository = "https://github.com/alessioarcara/rtmdet_pytorch_implementation.git"

[tool.flit.module]
name = "rtmdet"

[tool.mypy]
files = ["rtmdet"]
install_types = true
ignore_missing_imports = true
